% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abgr.R
\name{abgr}
\alias{abgr}
\title{Adaptive buffered grey  rolling model}
\usage{
abgr(y = y, ntest = NULL, model = gm, buff = operator, term = 1)
}
\arguments{
\item{y}{raw data for modelling}

\item{model}{unit grey model used in data piece}

\item{buff}{buffer operator used to data piece}

\item{term}{forecasting term by extroplation}
}
\description{
Generate ABGR model, computation results and its fitting graphes
}
\examples{
abgr(dataset = y, model = gm_1, buff = operator, term = 3)
}
