% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abrgm.R
\name{abrgm}
\alias{abrgm}
\title{Adaptive buffered grey rolling model
generate ABGR model, computation results and its fitting graphes}
\usage{
abrgm(y = y, ntest = NULL, model = gm, buff = operator, term = 1)
}
\arguments{
\item{y}{raw data for modelling}

\item{model}{unit grey model used in data piece}

\item{buff}{buffer operator used to data piece}

\item{term}{forecasting term by extroplation}
}
\description{
Adaptive buffered grey rolling model
generate ABGR model, computation results and its fitting graphes
}
\examples{
abrgm(dataset = y, model = gm_1, buff = operator, term = 3)
}
