% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background1.R
\name{background1}
\alias{background1}
\title{multi-projection backaground values
w:以0.5为基准的左右白化方程分割参数}
\usage{
background1(yd, w = 0.14)
}
\description{
multi-projection backaground values
w:以0.5为基准的左右白化方程分割参数
}
