% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boobj.R
\name{bo.obj}
\alias{bo.obj}
\title{Generate `bo` Object}
\usage{
bo.obj(x, formula = NULL, expression = NULL)
}
\arguments{
\item{x}{Raw data}

\item{formula}{Formula}

\item{expression}{Expression}
}
\description{
Generate `bo` Object
}
