% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialmethod.R
\name{buffplot}
\alias{buffplot}
\title{compare and plot the data graph of buffer operator effect
nr: number that displays arrows without}
\usage{
buffplot(x, buff = NULL, alpha = NA, add = FALSE, nr = 4)
}
\description{
compare and plot the data graph of buffer operator effect
nr: number that displays arrows without
}
