% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.greyforecasting}
\alias{coef.greyforecasting}
\title{fetch the parameters in grey forecasting object}
\usage{
\method{coef}{greyforecasting}(x)
}
\arguments{
\item{y}{data sequence}
}
\description{
fetch the parameters in grey forecasting object
}
\examples{
g<-gm(y)
coef(g)
}
