% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine}
\alias{combine}
\title{Combine multi-graphs based on matrix "layoutset"}
\usage{
combine(layoutset, ...)
}
\arguments{
\item{layoutset}{matrix "layoutset" may consist of dimensions or facets sequences}

\item{...}{#' @export}
}
\description{
Combine multi-graphs based on matrix "layoutset"
}
