% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareseq.r
\name{compareseq}
\alias{compareseq}
\title{compare errors according to data point's name
absolute errors:ae
percentage absolute errors:pae}
\usage{
compareseq(y1, y2)
}
\description{
compare errors according to data point's name
absolute errors:ae
percentage absolute errors:pae
}
