% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conplot.R
\name{conplot}
\alias{conplot}
\title{combination plot function}
\usage{
conplot(
  g,
  xlab = "x-axis",
  ylab = "y-axis",
  location = "topleft",
  legend = NA,
  forecast = FALSE
)
}
\arguments{
\item{g}{grey model list, a list variable containing models.}

\item{xlab}{label in x-axis.}

\item{ylab}{label in y-axis.}

\item{location}{legend location.}

\item{forecast}{present the forecasted values in plot.}
}
\description{
plot grey models in single graph
}
