% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm}
\alias{dgm}
\title{discrete grey prediction model by Xie Naiming}
\usage{
dgm(y, ntest = NULL, term = 1, buff = NULL, alpha = NA)
}
\arguments{
\item{y}{original data}

\item{ntest}{number of insample data to test accuracy}

\item{term}{length of extropotation data, forecasting data.}

\item{buff}{buffer operator used for original data.}
}
\description{
model sequential data with discrete GM(1,1) model
}
\examples{
g<-dgm(y)
}
