% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghg}
\alias{ghg}
\title{Energy-related greenhouse emission data from China Statistics Bureau}
\format{
data.frame variable containing the emission data from total emission, coal related source, oil related and gas related:
\describe{
    \item{year}{year, from 2000 to 2015}
    \item{Total}{total emission}
    \item{from Coal}{Coal related source emission}
    \item{from Oil}{Oil related source emission}
    \item{from Gas}{Gas related source emission}
}
}
\source{
\url{http://www.stats.gov.cn/tjsj/ndsj/2018/indexch.htm}
}
\usage{
ghg
}
\description{
GHG emission data from 2000-2015
}
\keyword{datasets}
