% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.R
\name{gm}
\alias{gm}
\title{classic grey forecasting model, GM(1,1)}
\usage{
gm(
  y,
  ntest = NULL,
  term = 1,
  present = c(NA, NA),
  bg = background,
  buff = NULL,
  alpha = NA,
  ...
)
}
\arguments{
\item{present}{character vector containing xlab and ylab.}

\item{bg}{background formula.}

\item{buff}{buffer operator used for original data.}

\item{alpha}{coefficient in buffer operator if used.}

\item{x}{data sequence.}
}
\description{
model sequential data with GM(1,1) model
}
\examples{
g<-gm(y)
}
