% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmn.R
\name{gmn}
\alias{gmn}
\title{GM(1,N)}
\usage{
gmn(
  y,
  xn,
  ntest = NULL,
  term = 1,
  bg = background,
  buff = NULL,
  alpha = NA,
  ...
)
}
\arguments{
\item{xn}{driven sequence.}

\item{bg}{background formula.}

\item{buff}{buffer operator used for original data.}

\item{alpha}{coefficient in buffer operator if used.}

\item{x}{data sequence.}
}
\description{
Grey prediction model with n input variables. uncompleted
}
\examples{
g<-gmn(y)
}
