% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmprocess.r
\name{gmprocess}
\alias{gmprocess}
\title{GM(1,1)模型完整处理过程
默认返回1期预测值，即term=1
修改pattern="model"，则返回包含所有参数的模型数据集}
\usage{
gmprocess(y, term = 1, pattern = "forecast")
}
\description{
GM(1,1)模型完整处理过程
默认返回1期预测值，即term=1
修改pattern="model"，则返回包含所有参数的模型数据集
}
