% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsplot.R
\name{gsplot}
\alias{gsplot}
\title{plot fitting graph of grey models based on ggplot2
plot multiple of single grey model}
\usage{
gsplot(
  ...,
  lastpoint = NA,
  xintercept = NA,
  legend = NA,
  save = F,
  n_breaks = 5
)
}
\arguments{
\item{lastpoint}{the cut line of x axis for fitting presentation.}

\item{xintercept}{divide line of insample part and forecasting part.}

\item{legend}{a charactor vector for model's name.}
}
\description{
plot fitting graph of grey models based on ggplot2
plot multiple of single grey model
}
