% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iago.R
\name{iago}
\alias{iago}
\title{IAGO operator}
\usage{
iago(x, r = 1)
}
\arguments{
\item{x}{sequential data}

\item{r}{order of accumulation operating generation.}
}
\description{
Inverse accumulation operating generation
}
