% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logiscost}
\alias{logiscost}
\title{Annual cost of Chinese food cold logistics demand}
\format{
An vector of 2014-2019
\describe{
    \item{logiscost}{total logistics, trillion yuan}
}
}
\source{
\url{China Cold Chain Logisitics Development Report(2020)}
}
\usage{
logiscost
}
\description{
A dataset of Chinese food cold logistics cost from 2014 to 2019, published by Cold Chain Logistics Commitee of CFLP
}
\keyword{datasets}
