% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logisquantity}
\alias{logisquantity}
\title{Annual quantity of Chinese food cold logistics demand}
\format{
An vector of 2014-2019
\describe{
    \item{logisquantity}{total quantity of cold logisticsdemand, trillion yuan}
}
}
\source{
\url{China Cold Chain Logisitics Development Report(2020)}
}
\usage{
logisquantity
}
\description{
A dataset of Chinese cold logistics cost from 2014 to 2019, published by Cold Chain Logistics Commitee of CFLP
}
\keyword{datasets}
