% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1.R
\name{mymodel}
\alias{mymodel}
\title{adaptive buffered grey  rolling model
usage:
mymodel(dataset=y,model=gm_1,buff=operator,term=3)}
\usage{
mymodel(dataset = y, model = gm, buff = operator, term = 3)
}
\description{
adaptive buffered grey  rolling model
usage:
mymodel(dataset=y,model=gm_1,buff=operator,term=3)
}
