% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{owners}
\alias{owners}
\title{Quantity of annual vehicle ownership}
\format{
An vector of 2006-2010
\describe{
    \item{p}{price, in dollars}
}
}
\source{
\url{https://data.stats.gov.cn/}
}
\usage{
owners
}
\description{
A dataset of Chinese vehicle owner
}
\keyword{datasets}
