% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgm.R
\name{pgm}
\alias{pgm}
\title{grey power model}
\usage{
pgm(y, r = 2, ntest = NULL, term = 1, bg = background, buff = NULL, alpha = NA)
}
\arguments{
\item{bg}{background formula.}

\item{buff}{buffer operator used for original data.}

\item{alpha}{coefficient in buffer operator if used.}

\item{x}{data sequence.}

\item{present}{character vector containing xlab and ylab.}
}
\description{
saturated growth, S shape tendency
}
\examples{
g<-pgm(y)
}
