% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.greyforecasting}
\alias{print.greyforecasting}
\title{Print grey forecasting model}
\usage{
\method{print}{greyforecasting}(x)
}
\arguments{
\item{x}{a greyforecasting object}
}
\description{
Print grey forecasting model
}
