% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project}
\alias{project}
\title{Progject file for grey prediction application}
\usage{
project(y, term = 1, ntest = NULL, mdlist = "gm")
}
\arguments{
\item{mdlist}{vector of character variable, which contains model names}
}
\description{
Progject file for grey prediction application
}
