% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{generate a grey rolling model
usage:
r<-roll(y)
arguments:
y: data sequence ; model: unit grey model used in each data piece
buff: the buffer operator function adjusting data piece
rollterm: forecasting terms by extroplation
piece: length of data piece ; stepsize: rolling step}
\usage{
roll(y, model = gm, buff = NA, intensity = NA, present = c(NA, NA),
  rollterm = 3, piece = 4, stepsize = 1)
}
\description{
generate a grey rolling model
usage:
r<-roll(y)
arguments:
y: data sequence ; model: unit grey model used in each data piece
buff: the buffer operator function adjusting data piece
rollterm: forecasting terms by extroplation
piece: length of data piece ; stepsize: rolling step
}
