% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smth.R
\name{smth}
\alias{smth}
\title{full information buffer operator}
\usage{
smth(x)
}
\arguments{
\item{y}{original data}

\item{r}{weight,lambda}

\item{w}{weight,gamma}
}
\description{
full information buffer operator
}
\examples{
smth(y,r=0.5,w=0.5)

}
