% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.greyforecasting}
\alias{summary.greyforecasting}
\title{Make the summary of greyforecasting object}
\usage{
\method{summary}{greyforecasting}(x)
}
\arguments{
\item{x}{a greyforecasting object}
}
\description{
Make the summary of greyforecasting object
}
\examples{
g<-gm(y)
summary(g)
}
