% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swvbo.r
\name{svwbo}
\alias{svwbo}
\title{smooth buffer operator with variable weight
y:original data
alpha:weight
format:y<-svwbo(y,alpha=0.5)
defaut alpha=0.5}
\usage{
svwbo(y, alpha = 0.5)
}
\description{
smooth buffer operator with variable weight
y:original data
alpha:weight
format:y<-svwbo(y,alpha=0.5)
defaut alpha=0.5
}
