% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verhulst.R
\name{verhulst}
\alias{verhulst}
\title{grey verhulst model}
\usage{
verhulst(y, ntest = NULL, term = 1, bg = background, buff = NULL, alpha = NA)
}
\arguments{
\item{bg}{background formula.}

\item{buff}{buffer operator used for original data.}

\item{alpha}{coefficient in buffer operator if used.}

\item{x}{data sequence.}

\item{present}{character vector containing xlab and ylab.}
}
\description{
saturated growth, S shape tendency
}
\examples{
g<-verhulst(y)
}
