% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y}
\alias{y}
\title{A dataset containing the emissions from China}
\format{
An atom vector from 2000-2014:
\describe{
    \item{year}{year, from 2000 to 2014}
}
}
\source{
\url{http://www.diamonds.info/}
}
\usage{
y
}
\description{
A dataset containing the emissions from China
}
\keyword{datasets}
