% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yr}
\alias{yr}
\title{Declining sequential test data
Test data generated from y by rev() function}
\format{
An atom vector from 2000-2014:
\describe{
    \item{year}{year, from 2000 to 2014}
}
}
\source{
\url{http://www.diamonds.info/}
}
\usage{
yr
}
\description{
Declining sequential test data
Test data generated from y by rev() function
}
\keyword{datasets}
