% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ys}
\alias{ys}
\title{Statistic data of annual accident death of China land traffic}
\format{
An vector of 1990-2003
\describe{
    \item{ys}{death number, 10000 persons}
}
}
\source{
\url{https://data.stats.gov.cn/}
}
\usage{
ys
}
\description{
A dataset of Chinese vehicle owner
}
\keyword{datasets}
