% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specialmethod.R
\name{buffplot}
\alias{buffplot}
\title{Compare and plot the data graph of buffer operator effect}
\usage{
buffplot(x, buff = NULL, alpha = NA, add = FALSE, nr = 4)
}
\arguments{
\item{nr}{}
}
\value{
nr: number that displays arrows without
}
\description{
Compare and plot the data graph of buffer operator effect
}
