% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.greyforecasting}
\alias{coef.greyforecasting}
\title{fetch the parameters in grey forecasting object
usage:
g<-gm(y)
coef(g)}
\usage{
\method{coef}{greyforecasting}(x)
}
\description{
fetch the parameters in grey forecasting object
usage:
g<-gm(y)
coef(g)
}
