% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.greyforecasting}
\alias{coef.greyforecasting}
\title{Fetch the parameters in grey forecasting object}
\usage{
\method{coef}{greyforecasting}(object)
}
\arguments{
\item{object}{a greyforecasting object}
}
\description{
Fetch the parameters in grey forecasting object
}
\examples{
g<-gm(y)

coef(g)

}
