% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.r
\name{combine}
\alias{combine}
\title{combine multi-graphs based on matrix "layoutset"
matrix "layoutset" may consist of dimensions or facets sequences}
\usage{
combine(layoutset, ...)
}
\description{
combine multi-graphs based on matrix "layoutset"
matrix "layoutset" may consist of dimensions or facets sequences
}
