% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareseq.r
\name{compareseq}
\alias{compareseq}
\title{Compare errors according to data point's name}
\usage{
compareseq(y1, y2)
}
\arguments{
\item{y1}{first data poinnt}

\item{y2}{second data point}
}
\value{
absolute errors:ae
percentage absolute errors:pae
}
\description{
Compare errors according to data point's name
}
