% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conplot.R
\name{conplot}
\alias{conplot}
\title{Combination plot function}
\usage{
conplot(g, xlab = NA, ylab = NA, legend = NA)
}
\arguments{
\item{g}{grey model, a list variable containing models}

\item{xlab}{label in x-axis}

\item{ylab}{label in y-axis}
}
\description{
plot grey models in single graph
}
