% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.r
\name{gm}
\alias{gm}
\title{Evaluate paramters of GM model
generate background formula, response formula and forecast values
formate: gdata<-gm(gdata)}
\usage{
gm(gdata)
}
\description{
Evaluate paramters of GM model
generate background formula, response formula and forecast values
formate: gdata<-gm(gdata)
}
