% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gm.r
\name{gm}
\alias{gm}
\title{classic grey forecasting model, GM(1,1)}
\usage{
gm(x, term = 1, present = c(NA, NA), bg = background, buff = NULL,
  alpha = NA, ...)
}
\arguments{
\item{x}{x: data sequence.}

\item{present}{present: character vector containing xlab and ylab.}

\item{bg}{bg: background formula.}

\item{buff}{buff: buffer operator used for original data.}

\item{alpha}{alpha: coefficient in buffer operator if used.}
}
\description{
model sequential data with GM(1,1) model
}
\examples{
g<-gm(y)
}
