% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmprocess.r
\name{gmprocess}
\alias{gmprocess}
\title{a integral GM(1,1) process using the tool functions
defaut forecasting term is 1, that is term=1
return a integral modelset when pattern="model"}
\usage{
gmprocess(y, pattern = "forecast")
}
\description{
a integral GM(1,1) process using the tool functions
defaut forecasting term is 1, that is term=1
return a integral modelset when pattern="model"
}
