% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1.R
\name{mymodel}
\alias{mymodel}
\title{adaptive buffered grey  rolling model}
\usage{
mymodel(dataset = y, model = gm, buff = operator, term = 3)
}
\arguments{
\item{dataset}{raw data for modelling}

\item{model}{unit grey model used in data piece}

\item{buff}{buffer operator used to data piece}

\item{term}{forecasting term by extroplation}
}
\description{
generate ABGR model, computation results and its fitting graphes
}
\examples{
mymodel(dataset=y,model=gm_1,buff=operator,term=3)
}
