% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.greyforecasting}
\alias{plot.greyforecasting}
\title{generate fitting graph of grey forecasting model
usage:
g<-gm(y)
plot(g)
plot(g,forecast=TRUE)}
\usage{
\method{plot}{greyforecasting}(x, location = "topleft", add = FALSE, forecast = FALSE)
}
\description{
generate fitting graph of grey forecasting model
usage:
g<-gm(y)
plot(g)
plot(g,forecast=TRUE)
}
