% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.greyforecasting}
\alias{plot.greyforecasting}
\title{generate fitting graph of grey forecasting model}
\usage{
\method{plot}{greyforecasting}(md, location = "topleft", forecast = FALSE)
}
\arguments{
\item{md}{'greyforecasting' object.}

\item{location}{legend position,'topleft','topright'.}

\item{forecast}{add the forecast data in plot.}
}
\description{
generate fitting graph of grey forecasting model
}
\examples{
g<-gm(y)
plot(g)
plot(g,forecast=TRUE)
}
