% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll.R
\name{roll}
\alias{roll}
\title{adaptive buffered grey rolling model}
\usage{
roll(y, ntest = NULL, rollterm = 1, model = gm, buff = NA,
  intensity = NA, present = c(NA, NA), piece = 4, stepsize = 1)
}
\arguments{
\item{y}{data sequence}

\item{rollterm}{forecasting terms by extroplation}

\item{model}{unit grey model used in each data piece}

\item{buff}{the buffer operator function adjusting data piece}

\item{piece}{length of data piece}

\item{stepsize}{rolling step}
}
\description{
adaptive buffered grey rolling model
}
\examples{
r<-roll(y)
}
