% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.greyforecasting}
\alias{summary.greyforecasting}
\title{make the summary of greyforecasting object
usage:
g<-gm(y)
summary(g)}
\usage{
\method{summary}{greyforecasting}(x)
}
\description{
make the summary of greyforecasting object
usage:
g<-gm(y)
summary(g)
}
