% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svwbo.R
\name{svwbo}
\alias{svwbo}
\title{smooth buffer operator with variable weight}
\usage{
svwbo(x, alpha = 0.5)
}
\arguments{
\item{alpha}{weight,alpha=0.5}

\item{y}{original data}
}
\description{
smooth buffer operator with variable weight
}
\examples{
format:y<-svwbo(y,alpha=0.5)
}
