% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{svwbo}
\alias{svwbo}
\title{smooth buffer operator with variable weight}
\usage{
svwbo(y, alpha = 0.5)
}
\arguments{
\item{y}{original data}

\item{alpha}{weight,alpha=0.5}
}
\description{
smooth buffer operator with variable weight
}
\examples{
format:y<-svwbo(y,alpha=0.5)

}
